ssh-known-hosts CLI should handle hashed hostnames
To store hashes instead of hostname or IP in clear text in the config,
update the script for ssh-known-hosts so that shell special characters
are correctly escaped rather than interpreted. For this, surround the
variable for the array of hosts by single quotes, which also requires
a change in how the array is read.

VRVDR-46681