ssh-known-hosts hostname or IP address should be hashed
The modelled copy command automatically adds ssh-known-hosts config
with hashed hostname/IP due to the fix for VRVDR-50920. So use of the
ssh-known-hosts generator config fetch-from-server and load-from-file
is no longer needed. Rather than deprecating these, fix these now by
using a hash in the config and also known hosts file. Note that there
is little benefit in modifying the ssh-known-hosts host config handling
itself to only hash hosts in the known hosts file, as the config is
accessible with less privileges. The host config itself is indifferent
as to whether host is in clear-text or hash, the former continuing to
be supported for reasons of backwards compatibility, the latter should
now be in dominant use.

VRVDR-46681