ssh-known-hosts: allow deletion using clear-text hostname/IP
The use of the copy command or ssh-known-hosts fetch-from-server config
results in the hash of the hostname or IP being used for the host in
the ssh-known-hosts config. Correspondingly, the host can only be
removed from the config by specifying the hash. A helper script is now
provided so that the host can be deleted using the clear-text hostname
or IP.

VRVDR-46681